Made a Secp ArtifactSignerFactory
