+++ core/src/main/java/tech/pegasys/eth2signer/core/multikey/metadata/HashicorpSigningMetadata.java
@@ -102,7 +102,7 @@ public class HashicorpSigningMetadata implements SigningMetadata {
   }
 
   @Override
-  public ArtifactSigner createSigner(final ArtifactSignerFactory factory) {
-    return factory.create(this);
+  public ArtifactSigner createSigner(final ArtifactSignerFactory artifactSignerFactory) {
+    return artifactSignerFactory.create(this);
   }
 }
