+++ core/src/test/java/tech/pegasys/eth2signer/core/multikey/BlsArtifactSignerFactoryTest.java
@@ -16,6 +16,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import tech.pegasys.eth2signer.core.multikey.metadata.ArtifactSignerFactory;
+import tech.pegasys.eth2signer.core.multikey.metadata.BlsArtifactSignerFactory;
 import tech.pegasys.eth2signer.core.multikey.metadata.FileKeyStoreMetadata;
 import tech.pegasys.eth2signer.core.multikey.metadata.HashicorpSigningMetadata;
 import tech.pegasys.eth2signer.core.multikey.metadata.SigningMetadataException;
@@ -34,7 +35,7 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.io.TempDir;
 
-class ArtifactSignerFactoryTest {
+class BlsArtifactSignerFactoryTest {
 
   private static final String PUBLIC_KEY =
       "95e57532ede3c1dd879061153f9cfdcdefa9dc5fb9c954a6677bc6641b8d26e39f70b660bbaa732c47277c0096e11400";
@@ -57,7 +58,7 @@ class ArtifactSignerFactoryTest {
     Files.copy(Path.of(Resources.getResource(PASSWORD_FILE).getPath()), passwordFile);
 
     artifactSignerFactory =
-        new ArtifactSignerFactory(
+        new BlsArtifactSignerFactory(
             configDir, new NoOpMetricsSystem(), new HashicorpConnectionFactory(vertx));
   }
